keep shared request preparation in steps
