--- pkg/coordinator/engine/engine.go
@@ -1,24 +1,0 @@
-/*
-Copyright 2026 The llm-d Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// Package engine defines values shared by coordinator engine implementations.
-package engine
-
-// Request carries a protocol path and body for the HTTP transport.
-type Request struct {
-	Path string
-	Body map[string]any
-}
