+++ pkg/coordinator/engine/vllm/decode.go
@@ -17,34 +17,21 @@ limitations under the License.
 package vllm
 
 import (
-	"maps"
-
 	reqcommon "github.com/llm-d/llm-d-router/pkg/common/request"
-	"github.com/llm-d/llm-d-router/pkg/coordinator/engine"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/gateway"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/pipeline"
 )
 
-// PrepareDecode mutates the terminal request body, including nested image parts.
-// Callers must not use those values concurrently.
-func (e Engine) PrepareDecode(reqCtx *pipeline.RequestContext, kvParams map[string]any) engine.Request {
-	e.injectUUIDs(reqCtx)
-
-	format := resolveFormat(e.useOpenAIFormat, reqCtx.OriginalPath)
+// PrepareDecode adds vLLM transfer and token fields, including nested image UUIDs.
+func PrepareDecode(reqCtx *pipeline.RequestContext, kvParams map[string]any, format gateway.RequestFormat) {
+	injectUUIDs(reqCtx)
 	switch format {
 	case gateway.FormatChatCompletions:
 		reqCtx.Body[reqcommon.FieldKVTransferParams] = kvParams
-		e.injectTokensField(reqCtx)
+		injectTokensField(reqCtx, reqCtx.Body)
 	case gateway.FormatCompletions:
 		reqCtx.Body[reqcommon.FieldKVTransferParams] = kvParams
-		if len(reqCtx.TokenIDs) > 0 {
-			reqCtx.Body["prompt"] = reqCtx.TokenIDs
-		}
 	case gateway.FormatGenerate:
-		// The /inference/v1/generate engine reads transfer params only from
-		// sampling_params.extra_args; a top-level kv_transfer_params is ignored,
-		// so the decode worker never pulls the prefill KV over NIXL. Merge into
-		// the client's sampling_params to preserve max_tokens and other fields.
 		sampling, ok := reqCtx.Body[reqcommon.FieldSamplingParams].(map[string]any)
 		if !ok {
 			sampling = map[string]any{}
@@ -52,47 +39,51 @@ func (e Engine) PrepareDecode(reqCtx *pipeline.RequestContext, kvParams map[stri
 		}
 		setGenerateTransferParams(sampling, kvParams, nil)
 	}
-	return engine.Request{Path: reqCtx.OriginalPath, Body: reqCtx.Body}
 }
 
-func (e Engine) injectTokensField(reqCtx *pipeline.RequestContext) {
-	tokens := map[string]any{
-		"token_ids": reqCtx.TokenIDs,
+// PrepareConditionalDecode adds rendered tokens to a vLLM chat cache probe.
+func PrepareConditionalDecode(reqCtx *pipeline.RequestContext, body map[string]any, format gateway.RequestFormat) {
+	if format == gateway.FormatChatCompletions && len(reqCtx.TokenIDs) > 0 {
+		injectTokensField(reqCtx, body)
 	}
+}
+
+func injectTokensField(reqCtx *pipeline.RequestContext, body map[string]any) {
+	tokens := map[string]any{"token_ids": reqCtx.TokenIDs}
 	if features := buildMMFeatures(reqCtx.MultimodalEntries, false); features != nil {
 		tokens["features"] = features
 	}
-	reqCtx.Body["tokens"] = tokens
+	body["tokens"] = tokens
 }
 
-func (e Engine) injectUUIDs(reqCtx *pipeline.RequestContext) {
-	for i, part := range reqcommon.ImageParts(reqCtx.Body) {
-		if i >= len(reqCtx.MultimodalEntries) {
-			break
-		}
-		part["uuid"] = reqCtx.MultimodalEntries[i].Hash
+func injectUUIDs(reqCtx *pipeline.RequestContext) {
+	messages, ok := reqCtx.Body["messages"].([]any)
+	if !ok {
+		return
 	}
-}
 
-// PrepareConditionalDecode builds a cache probe without mutating reqCtx.Body.
-func (e Engine) PrepareConditionalDecode(reqCtx *pipeline.RequestContext) engine.Request {
-	body := maps.Clone(reqCtx.Body)
-	format := resolveFormat(e.useOpenAIFormat, reqCtx.OriginalPath)
-	switch format {
-	case gateway.FormatChatCompletions:
-		if len(reqCtx.TokenIDs) > 0 {
-			tokens := map[string]any{
-				"token_ids": reqCtx.TokenIDs,
-			}
-			if features := buildMMFeatures(reqCtx.MultimodalEntries, false); features != nil {
-				tokens["features"] = features
-			}
-			body["tokens"] = tokens
+	hashIdx := 0
+	for _, msg := range messages {
+		msgMap, ok := msg.(map[string]any)
+		if !ok {
+			continue
 		}
-	case gateway.FormatCompletions:
-		if len(reqCtx.TokenIDs) > 0 {
-			body["prompt"] = reqCtx.TokenIDs
+		content, ok := msgMap["content"].([]any)
+		if !ok {
+			continue
+		}
+		for _, part := range content {
+			partMap, ok := part.(map[string]any)
+			if !ok {
+				continue
+			}
+			if partMap["type"] != "image_url" {
+				continue
+			}
+			if hashIdx < len(reqCtx.MultimodalEntries) {
+				partMap["uuid"] = reqCtx.MultimodalEntries[hashIdx].Hash
+				hashIdx++
+			}
 		}
 	}
-	return engine.Request{Path: reqCtx.OriginalPath, Body: body}
 }
